* fix(PIM-1234): add lambda inactive flag

* fix(PIM-1234): merge main

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge remote-tracking branch 'origin/main'

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* Merge branch 'main' of https://github.com/jensvogt/awsmock

* feat(1234-new-windows-port): new JSON classes

* fix(1375-lambda-invocation): add start/stop/restart/enable/disable ap…

* fix(1375-lambda-invocation): add start/stop/restart/enable/disable ap…

* fix(1375-lambda-invocation): add start/stop/restart/enable/disable ap…

* Merge branch 'main' into feat(1234-new-windows-port)

* fix(1375-lambda-invocation): add start/stop/restart/enable/disable ap…

* fix(1375-lambda-invocation): add start/stop/restart/enable/disable ap…

* Merge branch 'main' into feat(1234-new-windows-port)

+++ src/db/src/entity/lambda/Lambda.cpp
@@ -134,8 +134,8 @@ namespace AwsMock::Database::Entity::Lambda {
         Core::Bson::BsonUtils::SetLongValue(lambdaDoc, "containerSize", containerSize);
         Core::Bson::BsonUtils::SetDocumentValue(lambdaDoc, "tags", tagsDoc);
         Core::Bson::BsonUtils::SetDocumentValue(lambdaDoc, "ephemeralStorage", ephemeralStorageDoc);
-        Core::Bson::BsonUtils::SetIntValue(lambdaDoc, "timeout", timeout);
-        Core::Bson::BsonUtils::SetIntValue(lambdaDoc, "concurrency", concurrency);
+        Core::Bson::BsonUtils::SetLongValue(lambdaDoc, "timeout", timeout);
+        Core::Bson::BsonUtils::SetLongValue(lambdaDoc, "concurrency", concurrency);
         Core::Bson::BsonUtils::SetStringValue(lambdaDoc, "codeSha256", codeSha256);
         Core::Bson::BsonUtils::SetDocumentValue(lambdaDoc, "environment", varDoc);
         Core::Bson::BsonUtils::SetDocumentValue(lambdaDoc, "code", code.ToDocument());
